refactor(section-error): Split up exports;
To avoid the ESLint warning `react-refresh/only-export-components`.
