+++ client/src/pages/Configurations/ConfigBuild/Sections/index.tsx
@@ -18,7 +18,8 @@ import { Checkbox } from '@components/Checkbox';
 import { Switch } from './Switch';
 import { NarrativeSelect } from './NarrativeSelect';
 import { useSectionUpdater } from './useSectionUpdater';
-import { SectionErrorProvider, useSectionError } from './SectionErrorContext';
+import { SectionErrorProvider } from './SectionErrorProvider';
+import { useSectionError } from './useSectionError';
 import classNames from 'classnames';
 import { Field } from '@components/Field';
 import { Label } from '@components/Label';
