+++ client/src/pages/Configurations/ConfigBuild/Sections/useSectionError.ts
@@ -1,0 +1,11 @@
+import { useContext } from 'react';
+import { SectionErrorContext } from './SectionErrorContext';
+
+export function useSectionError() {
+  const ctx = useContext(SectionErrorContext);
+  if (!ctx)
+    throw new Error(
+      'useSectionError must be used within a SectionErrorProvider'
+    );
+  return ctx;
+}
